load data once, then filter to post details.
- No tests yet
- Just load the data with a service. Save it so that other
people can use the data so you don't have to make calls
all the time. You make the call every time the user
reloads the browser. You could also cache the result, depending
on what you need.

- add webpack to handle the bundling and asset serving in dev.

- organize the application by feature. Each folder in the client/app
is a feature and self-contained.

- The client is decoupled from the server (the backend). They are
in their own folders to emphasize on this fact.